Fixed one mysterious error
There was a runtime error when I tried to add a number to a pre-existing number.  When trying to write to the array in the object, I said I was trying to write to undefined, which is not true.  So I added some unrelated code and ran the program again.  Now it had no isues.  I removed the added bit(again, unrelated) and the program ran perfectly fine afterwars.  I didn't alter the code for adding the number at all, so I am not sure what happened.   Computers move in mysterious ways.